+++ api/model/user_balance_history.go
@@ -7,8 +7,8 @@ import (
 type UserBalanceHistory struct {
 	ID            uint64    `json:"id"`
 	UserBalanceID uint64    `json:"user_balance_id"`
-	BalanceBefore uint64    `json:"balance_before"`
-	BalanceAfter  uint64    `json:"balance_after"`
+	BalanceBefore int       `json:"balance_before"`
+	BalanceAfter  int       `json:"balance_after"`
 	Activity      string    `json:"activity"`
 	Type          string    `json:"type"`
 	IP            string    `json:"ip"`
